Using src in test and renaming index to typescript-lambda-api
