+++ tests/src/test-components/model/TestUser.ts
@@ -1,4 +1,4 @@
-import { Principal } from "../../../../dist/index"
+import { Principal } from "../../../../src/typescript-lambda-api"
 
 export class TestUser extends Principal {
     public constructor(name: string, public readonly roles: string[]) {
