+++ tests/src/test-controllers/RestrictedTestController.ts
@@ -1,6 +1,6 @@
 import { injectable } from "inversify"
 
-import { apiController, controllerRolesAllowed, Controller, GET } from "../../../dist/index"
+import { apiController, controllerRolesAllowed, Controller, GET } from "../../../src/typescript-lambda-api"
 
 @apiController("/test-restricted")
 @controllerRolesAllowed("SUPER_SPECIAL_USER", "SPECIAL_USER")
