+++ tests/src/test-controllers/TestController.ts
@@ -1,7 +1,7 @@
 import { injectable } from "inversify"
 import { Response, Request } from "lambda-api"
 
-import { apiController, controllerProduces, header, pathParam, queryParam, response, request, rolesAllowed, produces, principal, Controller, GET } from "../../../dist/index"
+import { apiController, controllerProduces, header, pathParam, queryParam, response, request, rolesAllowed, produces, principal, Controller, GET } from "../../../src/typescript-lambda-api"
 
 import { TestUser } from '../test-components/model/TestUser';
